Better support for single-line block components
If a block component is invoked on a single line, indentation gets
confused. For instance:

  {{#link-to 'some.route'}}Example <span>text</span>{{/link-to}}

The next line gets indented with one shiftwidth. This commit avoids the
issue by checking against another pattern.

It also renames the variables "lnum" to "plnum", and "line" to "pline"
for better clarity ("previous" line).